Added Links to README.md Added BERICHT.md for documentation moved Jest files to tests folder
