+++ BERICHT.md
@@ -1,0 +1,23 @@
+# Projektbericht für die Todo Anwendung
+Diese Datei enthält den Projektbericht für die Todo Anwendung. Der Bericht wird im Rahmen des Moduls `Softwarequalität` an der [Fachhoschule Südwestfalen](www.fh-swf.de) erstellt.
+
+## Gruppenmitglieder
+* Christian Peters
+* Kevin Hillebrand
+* Tim Tomczek
+
+## Vorgehensweise
+
+
+## Lösungen
+
+
+## Probleme
+
+
+## Ergebnisse der QS
+### Gefundene Bugs
+| Verhalten | Erwartetes Verhalten                                                              |
+| --- |-----------------------------------------------------------------------------------|
+| Das Löschen eines Todo lässt die Anwendung abstürzen | Das Löschen eines Todos sollte das Todo aus der Liste und der Datenbank entfernen |
+| Das Ändern des Namens eines Todos hängt den neuen Namen an den alten an | Der Name des Todos sollte überschrieben werden                                    |
