+++ README.md
@@ -1,3 +1,9 @@
 # ToDo-Anwendung
 
 Diese Anwendung dient als Beispiel für das [Modul Softwarequalität](https://elearning.fh-swf.de/course/view.php?id=19693).
+
+## Links
+* [Aufgabe](https://github.com/fhswf/softwarequalitaet/tree/main/Exercises/CI_ToDo)
+* [SonarQube](https://hopper.fh-swf.de/sonarqube/dashboard?id=peters_todo&codeScope=overall)
+* [GitHub Actions](https://github.com/Ranea-D/todo/actions)
+* [GitHub](https://github.com/Ranea-D/todo)
