+++ backend/tests/todo.test.js
@@ -1,6 +1,6 @@
 import request from 'supertest';
-import { app, server, db } from './index';
-import getKeycloakToken from './utils';
+import { app, server, db } from '../index.js';
+import getKeycloakToken from './utils.js';
 
 let token; // Speichert den abgerufenen JWT-Token
 
@@ -9,7 +9,7 @@ beforeAll(async () => {
 });
 
 describe('GET /todos (unautorisiert)', () => {
-    it('sollte einen 401-Fehler zurückgeben, wenn kein Token bereitgestellt wird', async () => {
+    it.only('sollte einen 401-Fehler zurückgeben, wenn kein Token bereitgestellt wird', async () => {
         const response = await request(app).get('/todos'); // Kein Authorization-Header
 
         expect(response.statusCode).toBe(401);
